requirements.txt
fastapi~=0.92.0
uvicorn~=0.20.0
Pillow~=9.4.0
pytesseract~=0.3.10
gTTS~=2.3.1
tensorflow~=2.11.0
pandas~=1.5.3
numpy~=1.23.5
scikit-learn~=1.2.1
joblib~=1.2.0
keras~=2.11.0
opencv-python~=4.5.4.60
easyocr~=1.6.2
paddleocr~=2.6.1.3
paddlepaddle~=2.4.2
